Added support for 128 bit ints

+++ include/xbinary.h
@@ -7,7 +7,7 @@
 #define rotl(x, y) ((x << y) | (x >> (sizeof(x) * 8 - y)))
 #define rotr(x, y) ((x >> y) | (x << (sizeof(x) * 8 - y)))
 
-#define __to_bin__(x)   uint64_t __x__ = x;                                         \
+#define __to_bin__(x)   __uint128_t __x__ = x;                                      \
                         char __arr__[sizeof(x) * 8] = {'\0'};                       \
                         int16_t __i__;                                              \
                         for (__i__ = sizeof(x) * 8 - 2; __i__ >= 0; --__i__) {      \
